# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/planets": {
            "get": {
                "description": "Get random planets list from 1-8",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "planets"
                ],
                "summary": "Get planets",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.V1GetAllPlanetsResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.PlanetItem"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/handler.V1GetAllPlanetsResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/planets/{id}": {
            "get": {
                "description": "Get planet by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "planets"
                ],
                "summary": "Get planet by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Planet ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.V1GetPlanetByIDResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.V1GetPlanetByIDResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/handler.V1GetPlanetByIDResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.ItemInfo": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                }
            }
        },
        "handler.V1GetAllPlanetsResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.PlanetItem"
                    }
                }
            }
        },
        "handler.V1GetPlanetByIDResponse": {
            "type": "object",
            "properties": {
                "geology": {
                    "$ref": "#/definitions/handler.ItemInfo"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "overview": {
                    "$ref": "#/definitions/handler.ItemInfo"
                },
                "radius": {
                    "type": "integer"
                },
                "revolution": {
                    "type": "integer"
                },
                "rotation": {
                    "type": "integer"
                },
                "structure": {
                    "$ref": "#/definitions/handler.ItemInfo"
                },
                "temperature": {
                    "type": "integer"
                }
            }
        },
        "model.PlanetItem": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}